Añadidas funcionalidades de Mapper

+++ Backend/src/main/java/com/calendario/trabajadores/services/user/UserService.java
@@ -94,16 +94,17 @@ public class UserService {
         return Optional.of(response);
     }
 
-    //Metodo para borrar un usuario
-    public Optional<UsuarioDTO> softDelete(Long id) {
+    //Metodo para desactivar un usuario
+    public Optional<CrearEditarUsuarioResponse> softDelete(Long id) {
         var usuario = userRepository.findById(id);
         if (usuario.isEmpty()) {
             return Optional.empty();
         }
         Usuario temp = usuario.get();
-        temp.activo = false;
-        userRepository.save(temp);
-        return Optional.of(new UsuarioDTO(temp.id, temp.nombre, temp.apellido1, temp.apellido2, temp.email, null, temp.activo));
+        temp.setActivo(false);
+        Usuario usuarioSaved = userRepository.save(temp);
+        var response = userMapper.userToCreateEditResponse(usuarioSaved);
+        return Optional.of(response);
     }
 
     //Metodo para reactivar a un usuario
@@ -120,40 +121,21 @@ public class UserService {
 
     //En lugar de tener varios metodos para encontrar según los parametros que tenga, creamos un método que
     //sea capaz de devolvernos una lista de usuarios según los parametros que le pasemos
-    public Optional<List<UsuarioDTO>> listar(Optional<Boolean> activo) {
-        if (activo.isEmpty()) { //si no se pasa el parametro activo, devolvemos todos los usuarios
+    public Optional<List<CrearEditarUsuarioResponse>> listar(Optional<Boolean> activo) {
+        if (activo.isEmpty()) {
+            //si no se pasa el parametro activo, devolvemos todos los usuarios
             var lista = userRepository.findAll();
-            return Optional.of(lista.stream().map(usuario -> new UsuarioDTO(
-                    usuario.id,
-                    usuario.nombre,
-                    usuario.apellido1,
-                    usuario.apellido2,
-                    usuario.email,
-                    null,
-                    usuario.activo
-            )).toList());
-        } else if (activo.get()) { //si se pasa el parametro activo y es true, devolvemos los usuarios activos
-            var lista = userRepository.findByActivo(true);
-            return Optional.of(lista.stream().map(usuario -> new UsuarioDTO(
-                    usuario.id,
-                    usuario.nombre,
-                    usuario.apellido1,
-                    usuario.apellido2,
-                    usuario.email,
-                    null,
-                    usuario.activo
-            )).toList());
+
+            var list = lista.stream().map(usuario -> userMapper.userToCreateEditResponse(usuario)).toList();
+            return Optional.of(list);
+
         } else {
-            var lista = userRepository.findByActivo(false); //si se pasa el parametro activo y es false, devolvemos los usuarios inactivos
-            return Optional.of(lista.stream().map(usuario -> new UsuarioDTO(
-                    usuario.id,
-                    usuario.nombre,
-                    usuario.apellido1,
-                    usuario.apellido2,
-                    usuario.email,
-                    null,
-                    usuario.activo
-            )).toList());
+            //si se pasa el parametro activo y es true o false, devolvemos los usuarios activos o inactivos respectivamente
+            var lista = userRepository.findByActivo(activo.get());
+            //var lista = userRepository.findByActivo(true);
+            //var lista = userRepository.findByActivo(false);
+            var list = lista.stream().map(usuario -> userMapper.userToCreateEditResponse(usuario)).toList();
+            return Optional.of(list);
         }
     }
 
@@ -164,20 +146,32 @@ public class UserService {
 
     }
 
-    //Metodo para borrar un usuario
-    public Optional<UsuarioDTO> borrar(Long id) {
+    //Metodo para borrar un usuario (IMPORTANTE: uso solo de admin!! no para usuarios!)
+    public Optional<CrearEditarUsuarioResponse> borrar(Long id, String email) {
         var usuario = userRepository.findById(id);
         if (usuario.isEmpty()) {
             return Optional.empty();
         }
+        if (!usuario.get().getEmail().equals(email)) {
+            return Optional.empty();
+        }
         userRepository.delete(usuario.get());
-        return Optional.of(new UsuarioDTO(
-                usuario.get().id,
-                usuario.get().nombre,
-                usuario.get().apellido1,
-                usuario.get().apellido2,
-                usuario.get().email,
-                null,
-                usuario.get().activo));
+        return Optional.of(userMapper.userToCreateEditResponse(usuario.get()));
+    }
+
+    public Optional<CrearEditarUsuarioResponse> getUsuario(String email) {
+        var usuario = userRepository.findUsuarioByEmail(email);
+        if (usuario.isEmpty()) {
+            return Optional.empty();
+        }
+        return Optional.of(userMapper.userToCreateEditResponse(usuario.get()));
+    }
+
+    public Optional<CrearEditarUsuarioResponse> getUsuario(Long id) {
+        var usuario = userRepository.findById(id);
+        if (usuario.isEmpty()) {
+            return Optional.empty();
+        }
+        return Optional.of(userMapper.userToCreateEditResponse(usuario.get()));
     }
 }
